requirements.txt
torch>=1.12.0
torchvision>=0.13.0
pyyaml
Pillow
numpy
